pyproject.toml
[tool.poetry]
name = "smm-bot"
version = "0.1.0"
description = ""
authors = ["Ilya Lopatin <mrmochin@yandex.ru>"]
readme = "README.md"
packages = [{include = "smm_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
